ai-powerpoint-generator/requirements.txt
streamlit==1.24.0
openai==0.27.8
python-pptx==0.6.21
requests==2.31.0
python-dotenv==1.0.0